+++ daemon/containerfs_linux.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/internal/mounttree"
-	"github.com/docker/docker/internal/unshare"
+	"github.com/docker/docker/daemon/internal/unshare"
 	"github.com/docker/docker/pkg/fileutils"
 	containertypes "github.com/moby/moby/api/types/container"
 )
