+++ daemon/libnetwork/osl/namespace_linux.go
@@ -14,11 +14,11 @@ import (
 	"syscall"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/unshare"
 	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	"github.com/docker/docker/daemon/libnetwork/ns"
 	"github.com/docker/docker/daemon/libnetwork/osl/kernel"
 	"github.com/docker/docker/daemon/libnetwork/types"
-	"github.com/docker/docker/internal/unshare"
 	"github.com/vishvananda/netlink"
 	"github.com/vishvananda/netlink/nl"
 	"github.com/vishvananda/netns"
